pagesize back to int (mistake on merging)

+++ Demos/Blazorise.Demo/Pages/Tests/DataGrid/DataGridPage.razor.cs
@@ -28,7 +28,7 @@ public partial class DataGridPage
 
     private DataGrid<Employee> dataGrid;
     public long currentPage { get; set; } = 1;
-    public long currentPageSize  { get; set; } = 5; 
+    public int currentPageSize  { get; set; } = 5; 
 
     private bool editable = true;
     private bool fixedHeader = false;
